+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- New functions `uef_time:add_seconds/1,2`
+- New functions `uef_time:add_minutes/1,2`
+- New functions `uef_time:add_hours/1,2`
+- New functions `uef_time:add_days/1,2`
+- New functions `uef_time:add_months/1,2`
+- New functions `uef_time:add_weeks/1,2`
+- New functions `uef_time:add_years/1,2`
+- New functions `uef_time:add_time/1,2`
+
 ## [1.3.0] - 2019-05-03
 
 ### Added